Pass on development instructions (#2)
Added instructions on how to obtain and load the stratum-bfrt docker image. Also, I moved instructions for setting `STRATUM_BF_DOCKER_FLAG` as an optional last step, as that was preventing the stratum container to start (it was looking to mount `abs-path-to-stratum_bfrt` which didn't exist on my machine).

Apart from this, instructions work well. I was able to run all tests (excluding multicast and packet I/0 for which we don't have support yet).