Fix IntProgrammable installing invalid sessions
bfrt returns error when trying to install clone groups for quad-pipe
ports on a dual-pipe chip. For now, we decide whether to install 4 or 2
sessions based on the pipeconf name. This is a temporary workaround, as
it prevents us from using chip-independent pipeconfs (#40). Ideally, we
should be able to use gNMI to retrieve the chip type or platform name.