Add support for loopback mode testing (#44)
This change adds support for loopback testing by means of two keyless
tables and a fake Ethernet header:

- Table FabricEgress.loopback_testing.enable can be used to flag
outgoing packets as loopback (note: ports must be configured in loopback
mode independently of this table)
-  When enabled, an additional Ethernet-like header is prepended with a
special ether type to signal the ingress pipe that the packet is
entering the pipe a second time.
- Table FabricIngress.loopback_testing.punt_to_cpu can be used punt all
loopback packets as P4RT packet-in

PTF tests for this change are broken on stratum-bfrt because of
stratum/stratum-bfrt#40. All tests pass on stratum-bf for all profiles.