Refactor support for loopback mode testing (#55)
The new design removes the need for tables to enable CPU port loopback by assuming that such behavior can only be initiated via a packet-out. For this reason, we add a new flag to the packet-out header to distinguish between direct packet-outs and those intended for CPU loopback.

Also, we simplify stripping GTPU headers for INT reports in the egress parser.