PacketOut based on L2 header (#40)
Makes the P4 program Tofino chip model independent by removing the dependency on the cpu port at compile time.
Instead this information is inserted at runtime by the controller into the switch_info table, and stored in the metadata for usage in the egress stages. To remove the dependency in the ingress parser (to detect PacketOuts from dataplane packets), we now use a special Ethertype in the serialized metadata header of PacketOuts.