10k scale (#199)
This PR increases the table sizes to accommodate the 10K UE scale target. The numbers are intentionally on the conservative side w.r.t. ASIC resource usage and are not pushing the limits. We aim for ~50% TCAM/SRAM/etc utilization per stage and do not want to spread tables over multiple stages yet. This should leave plenty of space for future changes and extensions.