Prevent FabricOptimizedFieldDetectorTest from modifying const tables (#214)
The SDE does not reject these requests, but instead crashes under the right circumstances (table with sliced match fields).

As a remedy and additional safeguard, P4 code can hide constant tables with the @hidden annotation from the control plane. After all, if the table is const, why should the controller need access to it?

We're also addressing this in Stratum, by rejecting such requests early.