Split is_int to is_int_report and is_int_recirc
We want to skip the queue report if it is an INT report apcket or an
INT packet which will be recirculated later.
But we don't want to skip the egress_vlan table if it is an INT report
packet since we may push or pop the VLAN tag for it.