Use queue id 1 for traffic and run all test params
To avoid reporting the INT mirrored or recirculated packet, we use queue id
1 for our traffic, and watch only queue 1.

This commit also moves the SlicingTest from slicing.py to fabric_test.py so
we can reuse the class and it's functions.

Also, apply all possible test params to veryfy if queue report works for
every type of packets.