[SDFAB-324] Port INT control app from ONOS (#317)
* [SDFAB-324] Port INT control app from ONOS

* remove unused APIs

* Add IntManagerTest

* cleanup

* do not remove entries from other device

* Add SrConfig listener and improve tests

It is possible that the SR device config loaded after device added or INT
config loaded. So we also need to install INT rules when we got SR device
config update/add event.

* Handle host events and cleanup

We need to handle host event for the collector host since
the collector host can be discovered later facter we set up everything.