Implement control traffic rate limit test (#319)
This PR adds a line rate test to verify that control traffic is rate limited to 60 Mbit/s.
Contrary to other tests with control traffic, this one is run without background traffic, as the only way to get a byte rate measurement is on the port level. Sending other traffic would falsify that reading. This is a limitation of the XL710 NICs.