Use flow report type instead of local report type
There are 3 INT report types: flow, queue, and drop
Currently we are using "local" a a report type which is not correct.
In this PR we rename all "local report type" to "flow report type"

Note that we will keep the "local report header" since it is the correct
header name based on INT spec.