[SDFAB-462] Add INT linerate traffic trace test to assess performance of egress drop report filter (#405)
This test is in the same suite as the other linerate tests replaying a datacenter traffic trace. This test configures the switch to drop packets in the egress pipeline, by simply only configuring the ingress VLAN table. All packets are thus dropped at the egress VLAN table.

The purpose of this test is to assess whether the egress drop report filter is performing to its expected standards. It's worth stating that this test will increase the Jenkins linerate test job runtime notably, since each test in the traffic trace suite takes about 10 minutes to complete individually (mostly because of TRex).