[SDFAB-881] Verify `SLICE_MOBILE` presence before adding it
The `init` method was not checking the presence of `SLICE_MOBILE`
before adding it and this was not allowing the end of the upf
programmable.

Note that `SlicingException` cannot be avoided as the `init` is
done without any cluster coordination. In general you should expect
to see in two instances one time the SlicingException. After that
everything will be fine.