[SDFAB-1100] Leverage in-order FlowRuleService APIs
FabricUpfProgrammable leverages the new APIs to guarantee
in-order processing of the requests coming from the north.

If batch APIs are not used there is no guarantee about the
processing order in the FlowRuleService. Instead, the `striped`
API allow the apps to signal a preference in the requests
processing which is used by the FlowRuleService to serialize
all the requests, having equal key, on the same executor.

Depends on https://gerrit.onosproject.org/c/onos/+/25423